Fix(CI): Update Snippy Docker image tag
The CI was failing with an "Unable to find image" error for the SNIPPY process.
This was due to an incorrect Docker image tag for snippy:4.6.0--hdfd78af_0.

This commit updates the Docker image tag in modules/local/snippy/main.nf
to snippy:4.6.0--hdfd78af_1, which is available in the biocontainers registry,
thus resolving the image pull error.